# docs/openapi-auth.yaml
openapi: 3.0.3
info:
  title: ThinkCraft Auth API
  version: 1.0.0
  description: 手机号+验证码登录/注册接口文档
servers:
  - url: http://localhost:3000
paths:
...
  /api/verification/send:
    post:
      summary: 发送验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 发送失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/verification/verify:
    post:
      summary: 验证验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: 验证成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SendCodeRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          example: '13800138000'
        type:
          type: string
          enum: [register, login, bind]
          default: register
    VerifyCodeRequest:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
          example: '13800138000'
        code:
          type: string
          example: '123456'
        type:
          type: string
          enum: [register, login, bind]
          default: register
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: -1
        error:
          type: string